changes account (#22)
Here is a mix of changes, on both client and admin interfaces, hopefully later we will have a more consistent commit log on that provide information about a. where is the change happening, b. what is changing, in the not common case that is something complex how is changing, remember that the commits can be done locally in case you are working with several solutions to a problem, probably stashin draft, and before sending it its way you can squash it into one commit with the description as described here. Also if you need to send partial commits for testing purposes you can add and "update" message that you know can be squash before accepting the PR